Time: 4 ms (93.2%), Space: 10 MB (99.1%) - LeetHub

+++ 0518-coin-change-ii/0518-coin-change-ii.cpp
@@ -10,7 +10,7 @@ public:
         // - Pros: 
         // - Cons: 
         // ================================
-        vector<long long> dp(amount + 1, 0);
+        vector<unsigned int> dp(amount + 1, 0);
         dp[0] = 1;
 
         for (int coin : coins) {
